+++ src/Runner/IO/Standard.php
@@ -8,19 +8,12 @@ use Innmind\BlackBox\Runner\IO;
 enum Standard implements IO
 {
     case output;
-    case error;
 
     #[\Override]
     public function __invoke(string $data): void
     {
-        match ($this) {
-            self::output => (static function(string $data) {
-                // use echo instead of \fwrite(\STDOUT) because it is less prone
-                // to crashing and being unable to display the data
-                echo $data;
-            })($data),
-            // TODO crash when stream no longer writable ?
-            self::error => \fwrite(\STDERR, $data),
-        };
+        // use echo instead of \fwrite(\STDOUT) because it is less prone
+        // to crashing and being unable to display the data
+        echo $data;
     }
 }